reestructuring project paths and files
