+++ config/actions.js
@@ -1,0 +1,1 @@
+export const APPLY = 'apply';
