+++ hooks/createActions.js
@@ -1,0 +1,5 @@
+import { APPLY } from '../config/actions';
+
+const createAction = (key, value) => ({ type: APPLY, key, value });
+
+export default createAction;
